Updated CostCell and CSV functions to write the the headers and labels to csv to fix the error of empty CSV.
I checked the CSV myself, and can see the
date is written on the first row though
the first column remains empty, which means I need
to check save_to_csv function.

Additionally, there's 2 problems right now:
 - several CSV files being saved when it should
be 1 CSV file with month and user name in it.
 - 3 folders being created, both empty when
 it should be 1 folder
 - Save File Dialogue needs to be fixed as to
have it save to a desired location and not
look for files to load into. Need to rewrite it.

Given the resolution, I'm upping te version to 0.09