Removed DataManager and moved it's functions to CSVManagement and UserManagement respectively to consolidate the functions.
Fixed up Main and CostCell to get the headers going
Did further consistency checks and reduced several
lines and removed a duplicate line.

Managed to get the CSV and JSON to save in relevant locations
next is to fix up the folders and have JSON save in same location
as csv.